+++ application/model/types.hpp
@@ -3,6 +3,6 @@
 
 using Matrix = std::vector< std::vector< std::array< unsigned, (unsigned)Parameters::Count > > >;
 
-using Weights = std::array<unsigned, (unsigned)Parameters::Count>;
+using Weights = std::array<double, (unsigned)Parameters::Count>;
 
 #endif // TYPES_HPP
